backend/requirements.txt
fastapi
uvicorn
python-dotenv
yfinance
pandas
numpy
scikit-learn
xgboost
lightgbm
statsmodels
joblib
